Reorg the kql parser so that all the code isn't in a single file.
