+++ rust/experimental/query_engine/kql-parser/src/lib.rs
@@ -1,4 +1,11 @@
 pub(crate) mod date_utils;
 pub(crate) mod kql_parser;
+pub(crate) mod logical_expressions;
+pub(crate) mod query_expression;
+pub(crate) mod scalar_expression;
+pub(crate) mod scalar_function_expressions;
+pub(crate) mod scalar_primitive_expressions;
+pub(crate) mod shared_expressions;
+pub(crate) mod tabular_expressions;
 
 pub use kql_parser::*;
